count_service/requirements.txt
grpcio
grpcio-tools
grpcio-reflection  
redis
protobuf
